# docs/swagger.yaml
basePath: /
definitions:
...
  models.ActualizarCantidadDTO:
    properties:
      cantidad:
        minimum: 1
        type: integer
    required:
    - cantidad
    type: object
  models.AgregarPrendaUniformeDTO:
    properties:
      cantidad:
        minimum: 1
        type: integer
      id_prenda:
        type: integer
    required:
    - cantidad
    - id_prenda
    type: object
...
  models.CreateUniformeDTO:
    properties:
      descripcion:
        type: string
      id_segmento:
        type: integer
      nombre_uniforme:
        type: string
      prendas:
        items:
          $ref: '#/definitions/models.PrendaUniformeCantidadDTO'
        minItems: 1
        type: array
    required:
    - id_segmento
    - nombre_uniforme
    - prendas
    type: object
  models.ErrorInfo:
    properties:
      code:
        type: string
      details: {}
    type: object
  models.PrendaUniformeCantidadDTO:
    properties:
      cantidad:
        minimum: 1
        type: integer
      id_prenda:
        type: integer
    required:
    - cantidad
    - id_prenda
    type: object
...
  models.StandardResponse:
    properties:
      data: {}
      error:
        $ref: '#/definitions/models.ErrorInfo'
      errors: {}
      message:
        type: string
      meta: {}
      success:
        type: boolean
    type: object
...
  models.UpdateUniformeDTO:
    properties:
      descripcion:
        type: string
      id_segmento:
        type: integer
      nombre_uniforme:
        type: string
    type: object
...
host: localhost:8084
info:
  contact:
    email: soporte@sistal.com
    name: Soporte API
  description: |-
    API del Microservicio de Catálogo de Prendas y Uniformes
    Gestiona el catálogo de prendas, uniformes, temporadas y datos maestros
  license:
    name: MIT
  title: MS-Catalogo API
  version: "1.0"
paths:
...
  /api/v1/uniformes:
    get:
      consumes:
      - application/json
      description: Obtiene lista de uniformes con paginación y filtros
      parameters:
      - default: 1
        description: Número de página
        in: query
        name: page
        type: integer
      - default: 20
        description: Límite por página
        in: query
        name: limit
        type: integer
      - description: Filtrar por segmento
        in: query
        name: id_segmento
        type: integer
      - description: Buscar en nombre o descripción
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.StandardResponse'
      summary: Listar Uniformes
      tags:
      - Uniformes
    post:
      consumes:
      - application/json
      description: Crea un nuevo conjunto de uniforme
      parameters:
      - description: Datos del uniforme
        in: body
        name: uniforme
        required: true
        schema:
          $ref: '#/definitions/models.CreateUniformeDTO'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.StandardResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.StandardResponse'
      summary: Crear Uniforme
      tags:
      - Uniformes
  /api/v1/uniformes/{id_uniforme}:
    get:
      consumes:
      - application/json
      description: Obtiene información detallada de un uniforme con todas sus prendas
      parameters:
      - description: ID del uniforme
        in: path
        name: id_uniforme
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.StandardResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.StandardResponse'
      summary: Obtener Uniforme por ID
      tags:
      - Uniformes
    put:
      consumes:
      - application/json
      description: Actualiza información básica del uniforme
      parameters:
      - description: ID del uniforme
        in: path
        name: id_uniforme
        required: true
        type: integer
      - description: Datos a actualizar
        in: body
        name: uniforme
        required: true
        schema:
          $ref: '#/definitions/models.UpdateUniformeDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.StandardResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.StandardResponse'
      summary: Actualizar Uniforme
      tags:
      - Uniformes
  /api/v1/uniformes/{id_uniforme}/prendas:
    post:
      consumes:
      - application/json
      description: Agrega una prenda a un uniforme existente
      parameters:
      - description: ID del uniforme
        in: path
        name: id_uniforme
        required: true
        type: integer
      - description: Prenda a agregar
        in: body
        name: prenda
        required: true
        schema:
          $ref: '#/definitions/models.AgregarPrendaUniformeDTO'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.StandardResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.StandardResponse'
      summary: Agregar Prenda a Uniforme
      tags:
      - Uniformes
  /api/v1/uniformes/{id_uniforme}/prendas/{id_prenda}:
    delete:
      consumes:
      - application/json
      description: Elimina una prenda de un uniforme
      parameters:
      - description: ID del uniforme
        in: path
        name: id_uniforme
        required: true
        type: integer
      - description: ID de la prenda
        in: path
        name: id_prenda
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.StandardResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.StandardResponse'
      summary: Eliminar Prenda de Uniforme
      tags:
      - Uniformes
    put:
      consumes:
      - application/json
      description: Actualiza la cantidad de una prenda en el uniforme
      parameters:
      - description: ID del uniforme
        in: path
        name: id_uniforme
        required: true
        type: integer
      - description: ID de la prenda
        in: path
        name: id_prenda
        required: true
        type: integer
      - description: Nueva cantidad
        in: body
        name: cantidad
        required: true
        schema:
          $ref: '#/definitions/models.ActualizarCantidadDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.StandardResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.StandardResponse'
      summary: Actualizar Cantidad de Prenda en Uniforme
      tags:
      - Uniformes
  /garments/{id}:
    get:
      consumes:
      - application/json
      description: Get garment details by ID (legacy endpoint)
      parameters:
      - description: Garment ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get Garment (Legacy)
      tags:
      - Garments (Legacy)
...
schemes:
- http
- https
swagger: "2.0"
...